* Initial prototype build of Eyelink-Toolbox for M$-Windows:
  - Small modifications and cleanups to C source files to make
    them build cleanly on M$-Windows.

  - windowsmakeeyelink.m in PsychSourceGL/Source is the "Makefile"
    to build the ETB on Windows.

  - You need to have the Eyelink SDK for Windows installed. This is
    freely available (after free registration) from SR-Research.

-> Eyelink('Version') does not yet work. Does it on OS-X?

-> Quicktime SDK needs to be installed for no good reason. There's
   a bit of cleanup needed in Windows PTB-3 to resolve this.

-> Compiles cleanly, but only tested in dummy mode due to lack of
   a real Eyelink-Tracker.

--> Psychhardware/Eyelink/EyelinkBasic/Eyelink.dll is the mex file.

--> Should not interfere with OS-X build, so Safe for Beta.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@516 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Common/Eyelink/EyelinkDriftCorrStart.c
@@ -60,7 +60,7 @@ PsychError EyelinkDriftCorrStart(void)
    PsychCopyInIntegerArg(1, TRUE, &x);
    PsychCopyInIntegerArg(2, TRUE, &y);   
    
-   status = eyelink_driftcorr_start(x, y);
+   status = (int) eyelink_driftcorr_start((INT16) x,(INT16) y);
    
    /* if there is an output variable available, assign openstatus to it.   */
    PsychCopyOutDoubleArg(1, FALSE, status);
